# docs/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Agenda Barbearia API
  version: 1.1.0
  summary: Integrações de agenda (n8n, WhatsApp e app do cliente)
  description: |
    API REST para **conectar o WhatsApp, o n8n ou o app do cliente** à agenda da
    barbearia — as mesmas regras do site `/agenda`.

    **O que dá para fazer**
    - Ver dados da **loja** (`/shop`)
    - Ver **serviços** e **profissionais**
    - Consultar **horários livres**, **criar**, **listar**, **remarcar** e **cancelar**
    - Buscar **cliente** pelo WhatsApp
    - No **app**: ver e **editar o próprio cadastro** (`GET` / `PATCH /customers/me`),
      registrar push (`/customers/me/push-token`), ver notificações (`/customers/me/notifications`)
      e **sair de todos os aparelhos** (`DELETE /customers/me/sessions`)
    - Disparar e confirmar **lembretes** (~1h antes)
    - No **n8n**: ativar/pausar a **IA** por conversa (`/ai-status`)

    Comandas, caixa e comissões ficam **só no painel** — não entram nesta API.

    ## Repetir sem duplicar (Idempotency-Key)

    Em `POST /appointments`, `PATCH /appointments/{id}` e `DELETE /appointments/{id}`,
    envie um header `Idempotency-Key` (qualquer string única por tentativa,
    ex.: um UUID). Se a chamada for repetida com a mesma chave — por exemplo,
    um retry automático depois de um timeout — o servidor devolve a mesma
    resposta da primeira vez em vez de repetir a ação.

    ## Autenticação

    ### Chave de API (n8n / integrações)

...

    ```
...
    ```

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /appointments/{id}:
    parameters:
      - $ref: "#/components/parameters/appointmentId"
    patch:
      tags: [Agendamentos]
      operationId: updateAppointment
      summary: Remarcar agendamento
      description: |
        Troca data, horário, barbeiro e/ou serviços de um agendamento já existente.
        O WhatsApp do body precisa ser o do dono do horário.

        **Acesso:** privada. Scope `appointments:update`
        (chave de API, dono, cookie OTP ou `Bearer` com `accessToken`).
        WhatsApp do body deve ser o da sessão do cliente.

        Recomendado enviar `Idempotency-Key` (ver parâmetro) pra evitar
        remarcar em duplicidade num retry.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [whatsapp, professionalId, date, startTime, serviceIds]
              properties:
                whatsapp:
                  type: string
                professionalId:
                  type: string
                  format: uuid
                date:
                  type: string
                  format: date
                startTime:
                  type: string
                  example: "15:00"
                serviceIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 1
      responses:
        "200":
          description: Remarcado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  appointmentId:
                    type: string
                    format: uuid
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    delete:
      tags: [Agendamentos]
      operationId: cancelAppointment
      summary: Cancelar agendamento
      description: |
        Cancela o horário do cliente. O registro continua no sistema com status
        `cancelled` (não apaga o histórico).

        **Acesso:** privada. Scope `appointments:cancel`
        (chave de API, dono, cookie OTP ou `Bearer` com `accessToken`).
        WhatsApp da query deve ser o da sessão do cliente.

        Recomendado enviar `Idempotency-Key` (ver parâmetro) pra evitar
        cancelar em duplicidade num retry.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/whatsappQuery"
        - $ref: "#/components/parameters/idempotencyKeyHeader"
      responses:
        "200":
          description: Cancelado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  appointmentId:
                    type: string
                    format: uuid
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /ai-status:
    get:
      tags: [IA]
      operationId: getAiStatus
      summary: Consultar se a IA está ativa numa conversa
      description: |
        Devolve se o atendimento por IA está ativo ou pausado naquele
        WhatsApp. Sem registro salvo, a IA é considerada ativa (`true`).

        Pensada pra substituir o acesso direto do n8n ao banco (Postgres)
        por uma chave de API com escopo próprio.

        **Acesso:** privada. Scope `ai_status:read`.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/whatsappQuery"
      responses:
        "200":
          description: Status da IA na conversa
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  whatsapp:
                    type: string
                  iaAtiva:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    patch:
      tags: [IA]
      operationId: setAiStatus
      summary: Ativar ou pausar a IA numa conversa
      description: |
        Liga ou pausa o atendimento por IA num WhatsApp — por exemplo,
        quando um humano assume a conversa e a IA deve parar de responder.

        **Acesso:** privada. Scope `ai_status:write`.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [whatsapp, iaAtiva]
              properties:
                whatsapp:
                  type: string
                iaAtiva:
                  type: boolean
            example:
              whatsapp: "11999999999"
              iaAtiva: false
      responses:
        "200":
          description: Status atualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  whatsapp:
                    type: string
                  iaAtiva:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: dbc_live_… ou accessToken do OTP
      description: |
        Dois tipos de Bearer:
        1. **Chave de API** (`dbc_live_<keyId>_<secret>`) — n8n / integrações
        2. **Sessão do cliente** (`accessToken` de `POST /api/agenda/otp/verify`) — app mobile

        Exemplo: `Authorization: Bearer <token>`

        `/customers/me` aceita **somente** o token de sessão do cliente (não chave de API).
  parameters:
    whatsappQuery:
      name: whatsapp
      in: query
      required: true
      schema:
        type: string
      example: "11999999999"
      description: WhatsApp com DDD (com ou sem 55)
    appointmentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    idempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Chave opcional escolhida por quem chama (ex.: um UUID gerado antes
        de enviar). Se a mesma chamada for repetida com a mesma chave — por
        exemplo, um retry por timeout de rede do app ou do n8n — o servidor
        devolve a resposta salva da primeira vez, em vez de repetir a ação
        (evita agendamento duplicado). Reusar a chave com dados diferentes
        dá **409**.
      schema:
        type: string
        maxLength: 200
  schemas:
    Error:
      type: object
      description: |
        Formato de erro padrão de toda a API: `ok` sempre `false` e `error`
        com uma mensagem em português pronta pra mostrar pro usuário.
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: string
...
  responses:
    BadRequest:
      description: Parâmetros inválidos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Não autorizado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            ok: false
            error: Não autorizado.
    Forbidden:
      description: Sem permissão
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            ok: false
            error: Sem permissão.
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflito de regra de negócio
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limit excedido
      headers:
        Retry-After:
          schema:
            type: integer
          description: Segundos até tentar de novo
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Muitas tentativas em pouco tempo. Aguarde alguns minutos e tente de novo.
...