# src/app/api/v1/appointments/[id]/route.ts
import { NextRequest, NextResponse } from "next/server";
import { revalidatePath } from "next/cache";
import { z } from "zod";
import { safeApiRoute } from "@/lib/api/safe-route";
import { withProtectedApiRouteGuard } from "@/lib/api/with-api-guard";
import { withIdempotency } from "@/lib/api/idempotency";
import {
  cancelPublicAppointment,
  updatePublicAppointment,
} from "@/lib/manage-public-appointment";
import { protectedAuthRateLimitKey } from "@/lib/protected-api-auth";
import {
  normalizeWhatsapp,
  WHATSAPP_INVALID_MESSAGE,
  whatsappSchema,
} from "@/lib/whatsapp";

const updateBodySchema = z.object({
  whatsapp: whatsappSchema,
  professionalId: z.uuid(),
  date: z.string().regex(/^\d{4}-\d{2}-\d{2}$/),
  startTime: z.string().regex(/^([01]\d|2[0-3]):[0-5]\d$/),
  serviceIds: z.array(z.uuid()).min(1),
});

type RouteContext = { params: Promise<{ id: string }> };

// PATCH /api/v1/appointments/:id — remarcar agendamento do cliente
export async function PATCH(request: NextRequest, context: RouteContext) {
  return safeApiRoute(async () => {
    const { id } = await context.params;

    let json: unknown;
    try {
      json = await request.json();
    } catch {
      return NextResponse.json(
        { ok: false, error: "Corpo da requisição inválido." },
        { status: 400 }
      );
    }

    if (typeof json !== "object" || json === null) {
      return NextResponse.json(
        { ok: false, error: "Corpo da requisição inválido." },
        { status: 400 }
      );
    }

    const raw = json as Record<string, unknown>;
    const whatsapp =
      typeof raw.whatsapp === "string"
        ? normalizeWhatsapp(raw.whatsapp)
        : null;
    if (!whatsapp) {
      return NextResponse.json(
        { ok: false, error: WHATSAPP_INVALID_MESSAGE },
        { status: 400 }
      );
    }

    const parsed = updateBodySchema.safeParse({ ...raw, whatsapp });
    if (!parsed.success) {
      return NextResponse.json(
        { ok: false, error: parsed.error.issues[0].message },
        { status: 400 }
      );
    }

    return withProtectedApiRouteGuard(
      request,
      {
        scope: "appointments:update",
        rateLimit: "appointmentMutate",
        whatsapp,
      },
      async ({ auth }) =>
        withIdempotency(
          request,
          {
            route: "appointments.update",
            authIdentifier: protectedAuthRateLimitKey(auth) ?? "anonymous",
            resourceId: id,
            requestPayload: parsed.data,
          },
          async () => {
            const result = await updatePublicAppointment(id, parsed.data);

            if (!result.ok) {
              return NextResponse.json(
                { ok: false, error: result.error },
                { status: result.status }
              );
            }

            revalidatePath("/admin");
            revalidatePath("/agenda");
            return NextResponse.json({
              ok: true,
              appointmentId: result.data.id,
            });
          }
        )
    );
  });
}

// DELETE /api/v1/appointments/:id?whatsapp=... — cancelar agendamento do cliente
export async function DELETE(request: NextRequest, context: RouteContext) {
  return safeApiRoute(async () => {
    const { id } = await context.params;
    const raw = request.nextUrl.searchParams.get("whatsapp") ?? "";
    const whatsapp = normalizeWhatsapp(raw);
    if (!whatsapp) {
      return NextResponse.json(
        { ok: false, error: WHATSAPP_INVALID_MESSAGE },
        { status: 400 }
      );
    }

    return withProtectedApiRouteGuard(
      request,
      {
        scope: "appointments:cancel",
        rateLimit: "appointmentMutate",
        whatsapp,
      },
      async ({ auth }) =>
        withIdempotency(
          request,
          {
            route: "appointments.cancel",
            authIdentifier: protectedAuthRateLimitKey(auth) ?? "anonymous",
            resourceId: id,
            requestPayload: { whatsapp },
          },
          async () => {
            const result = await cancelPublicAppointment(id, whatsapp);

            if (!result.ok) {
              return NextResponse.json(
                { ok: false, error: result.error },
                { status: result.status }
              );
            }

            revalidatePath("/admin");
            revalidatePath("/agenda");
            return NextResponse.json({
              ok: true,
              appointmentId: result.data.id,
            });
          }
        )
    );
  });
}